fixed upgrade script and VLAN reset
git-svn-id: svn://svn.code.sf.net/p/phpipam/code@121 d9489115-519a-4661-88eb-0d2216f09706

+++ db/SCHEMA.sql
@@ -135,7 +135,7 @@ UNLOCK TABLES;
 CREATE TABLE `vlans` (
     `vlanId` INTEGER(11) NOT NULL AUTO_INCREMENT,
     `name` VARCHAR(255) CHARACTER SET utf8 COLLATE utf8_general_ci NOT NULL,
-    `number` INTEGER(3),
+    `number` INTEGER(4),
     `description` TEXT CHARACTER SET utf8 COLLATE utf8_general_ci,
     PRIMARY KEY (`vlanId`)
 ) ENGINE=MYISAM DEFAULT CHARSET=utf8;
