+++ db/UPDATE-v0.4.sql
@@ -25,7 +25,7 @@ ALTER TABLE `ipaddresses` ADD `mac` varchar(20) DEFAULT NULL after `dns_name`;	/
 CREATE TABLE `vlans` (
     `vlanId` INTEGER(11) NOT NULL AUTO_INCREMENT,
     `name` VARCHAR(255) CHARACTER SET utf8 COLLATE utf8_general_ci NOT NULL,
-    `number` INTEGER(3),
+    `number` INTEGER(4),
     `description` TEXT CHARACTER SET utf8 COLLATE utf8_general_ci,
     PRIMARY KEY (`vlanId`)
 ) ENGINE = MYISAM DEFAULT CHARSET=utf8;
